moved components to root folder
