+++ src/WireChatServiceProvider.php
@@ -10,8 +10,8 @@ use Namu\WireChat\Console\Commands\InstallWireChat;
 use Namu\WireChat\Facades\WireChat as FacadesWireChat;
 use Namu\WireChat\Livewire\Chat\Chat;
 use Namu\WireChat\Livewire\Chats\Chats;
-use Namu\WireChat\Livewire\Components\NewChat;
-use Namu\WireChat\Livewire\Components\NewGroup;
+use Namu\WireChat\Livewire\NewChat;
+use Namu\WireChat\Livewire\NewGroup;
 use Namu\WireChat\Livewire\Chat\Group\Permissions;
 use Namu\WireChat\Livewire\Index;
 use Namu\WireChat\Livewire\Chat\Group\AddMembers;
@@ -20,7 +20,7 @@ use Namu\WireChat\Livewire\Chat\Group\Members;
 use Namu\WireChat\Livewire\Modals\ChatDialog;
 use Namu\WireChat\Livewire\Chat\Drawer;
 use Namu\WireChat\Livewire\View;
-use Namu\WireChat\Livewire\Components\WireChat;
+use Namu\WireChat\Livewire\WireChat;
 use Namu\WireChat\Middleware\BelongsToConversation;
 use Namu\WireChat\Services\WireChatService;
